For async events, threads should write to pfw file while holding a lock to avoid missing entries.
